change(app): remove bsheet on home screen
